refactor(cli): split command handlers
Keep the public main(argv) driver in cli.ts while moving list, search, compare, model commands, and shared manager loading into focused CLI modules.

Closes #134
